Create README.md
Dataset obtained from the following websites: 

https://archive.ics.uci.edu/ml/datasets/Heart+Disease

https://www.kaggle.com/ronitf/heart-disease-uci#__sid=js0

Purpose of the Machine Learning exercise is to be able to predict based on different factors if a patient has a narrowing of < or >50%. 

Different machine models are explored logistic regression, random forest classification and support vector classifier. 

Notebook originally created with GoogleColab, then downloaded as a  jupyter notebook to local machine and edited as required.